Fix tests to support their execution in two different modes
By default, Python gets tests from its `PYTHONPATH`.
The line which was removed now in each test forced to get tests
from the source code directory for developing purposes.
