+++ tests/test_utils.py
@@ -24,13 +24,9 @@
 import copy
 import datetime
 import json
-import sys
 import threading
 import time
 import unittest
-
-if '..' not in sys.path:
-    sys.path.insert(0, '..')
 
 from arthur.utils import RWLock, JSONEncoder
